Extract the guarded image pipeline into src/api/substack/image.js
update_draft needs the same download-and-encode path for cover_image, and a
second copy of an SSRF guard is a second place to get it wrong. The error
prefix becomes a neutral `image:` so update_draft does not report failures
signed by a tool the caller never invoked.
